backend/requirements.txt
fastapi
uvicorn
pydantic>=2.9.0
pydantic-settings
supabase
python-dotenv
google-generativeai
celery
redis
PyPDF2
python-multipart
